+++ priv/repo/migrations/20201214072008_add_theme_pref_to_users.exs
@@ -3,7 +3,7 @@ defmodule Plausible.Repo.Migrations.AddThemePrefToUsers do
 
   def change do
     alter table(:users) do
-      add_if_not_exists :theme, :string, default: "system"
+      add_if_not_exists(:theme, :string, default: "system")
     end
   end
 end